requirements.txt
fastapi
uvicorn
twilio
google-generativeai
python-multipart
pydantic
python-dotenv
langchain
langchain-google-genai
sendgrid
certifi
pyngrok
requests
django
SQLAlchemy
mysql-connector-python